requirements.txt
lupa
PySide6
PyYAML
pyinstaller
pydantic
loguru